crawler-system/requirements.txt
# HTTP & Parsing
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
markdownify==0.12.1

# Playwright for JS-rendered sites
playwright==1.42.0

# Config
pyyaml==6.0.1

# Utils
tqdm==4.66.2
python-slugify==8.0.4

# Git operations
GitPython==3.1.42
